user modify scope set

+++ public/main.js
@@ -4,6 +4,8 @@ $('#update-req').click(function (e) {
     window.location.href = "http://localhost:3000/user-update.html"
 })
 
+
+//fuction for user to register on the platform//
 $(document).ready(function () {
     $('#register-btn').click(function (e) {
         e.preventDefault();
@@ -57,19 +59,20 @@ $(document).ready(function () {
                 }
             })
         }
-
-
     })
 })
 
 $('#id').text(`${sessionStorage.getItem('user')}`)
 
+
+// fuction for user to logoff the app
 $('#logout').click(function (e) {
     e.preventDefault();
     sessionStorage.clear();
     window.location.href = "http://localhost:3000/main.html"
 })
 
+//function for user login
 $(document).ready(function () {
     $('#login-btn').click(function (e) {
         e.preventDefault();
@@ -112,7 +115,7 @@ $(document).ready(function () {
     })
 })
 
-
+//function for user to request loan
 $(window).ready(function () {
     $('#request-btn').click(function (e) {
         e.preventDefault();
@@ -167,7 +170,7 @@ $(window).ready(function () {
     })
 })
 
-
+//function for user to request loans
 $(window).ready(function () {
     $("#check-requests").click(function (e) {
         e.preventDefault();
@@ -195,7 +198,7 @@ $(window).ready(function () {
     })
 })
 
-
+//function for user to modify their requests
 const modifyUserRequest = (ids, jsonData) => {
     $.ajax({
         type: "PATCH",
@@ -209,6 +212,7 @@ const modifyUserRequest = (ids, jsonData) => {
     })
 }
 
+//function for admin to approve user requests
 $('#approve').click(function (e) {
     e.preventDefault();
     let ids = $('#cusId').val();
@@ -219,7 +223,7 @@ $('#approve').click(function (e) {
     alert(`Request ${ids} has successfully been approved`);
 })
 
-
+//function for admin to decline user requests
 $('#loanDecline').click(function (e) {
     e.preventDefault();
     let ids = $('#cusId').val();
@@ -230,8 +234,8 @@ $('#loanDecline').click(function (e) {
     alert(`Request ${ids} has successfully been declined`);
 })
 
-
-
+let myId = []
+//function for user to get thier requests
 $(window).ready(function () {
     $("#user-requests").click(function (e) {
         e.preventDefault();
@@ -242,20 +246,24 @@ $(window).ready(function () {
             contentType: "application/json",
             success: function (data) {
                 for (let i = 0; i < data.length; i += 1) {
-                    if (data[i].username.toString() == sessionStorage.getItem('user').toString()) {
+
+                    if (data[i].userId == sessionStorage.getItem('id')) {
                         if (data[i].approval === true) {
+                            myId.push(data[i].id)
                             $("#dash-request").append(`<tr>
                             <td id="table-duration">${data[i].id}</td>
                             <td id="table-amount">${data[i].amount}</td>
                             <td id="table-amount">Request Approved</td> 
                           </tr>`)
                         } else if (data[i].approval === false) {
+                            myId.push(data[i].id)
                             $("#dash-request").append(`<tr>
                             <td id="table-duration">${data[i].id}</td>
                             <td id="table-amount">${data[i].amount}</td>
                             <td id="table-amount">Request Declined</td> 
                           </tr>`)
                         } else {
+                            myId.push(data[i].id)
                             $("#dash-request").append(`<tr>
                             <td id="table-duration">${data[i].id}</td>
                             <td id="table-amount">${data[i].amount}</td>
@@ -273,7 +281,7 @@ $(window).ready(function () {
 })
 
 
-
+//function for users to modify their requests
 $('#user-modify').click(function (e) {
     e.preventDefault();
     $('#user-modify').remove();
@@ -281,9 +289,9 @@ $('#user-modify').click(function (e) {
     $('#hidden').removeClass('hidded');
 })
 
-$('#user-delete').click(function (e) {
-    e.preventDefault();
-})
+// $('#user-delete').click(function (e) {
+//     e.preventDefault();
+// })
 
 $(window).ready(function () {
     $("#user-delete").click(function (e) {
@@ -295,33 +303,37 @@ $(window).ready(function () {
             dataType: "json",
             contentType: "application/json",
             success: function (data) {
-                for (let i = 0; i < data.length; i += 1) {
-                    if (data[i].id.toString() == ids.toString() && data[i].approval == true) {
-                        alert('Request has already been approved')
-                        i = data.length;
-                    } else if (data[i].id.toString() == ids.toString() && data[i].approval == false) {
+                if (myId.includes(ids)) {
+                    for (let i = 0; i < data.length; i += 1) {
 
-                        let jsonD = {
-
-                            "id": `${ids}`
-                        }
-                        $.ajax({
-                            url: `http://localhost:3000/borrowers/${ids}`,
-                            method: 'DELETE',
-                            dataType: 'json',
-                            contentType: 'application/json',
-                            data: JSON.stringify(jsonD),
-
-                            success: function (data) {
-                                console.log(data);
-                                alert('successfully DELETED');
-                                location.reload();
-                            },
-                            error: function () {
-                                alert('error')
+                        if (data[i].id.toString() == ids.toString() && data[i].approval == true) {
+                            alert('OOops!!! This request has been approved')
+                            i = data.length;
+                        } else if (data[i].id.toString() == ids.toString() && data[i].userId == sessionStorage.getItem('id') && data[i].approval == false) {
+                            let jsonD = {
+                                "id": `${ids}`
                             }
-                        })
+
+                            $.ajax({
+                                url: `http://localhost:3000/borrowers/${ids}`,
+                                method: 'DELETE',
+                                dataType: 'json',
+                                contentType: 'application/json',
+                                data: JSON.stringify(jsonD),
+
+                                success: function (data) {
+                                    console.log(data);
+                                    alert('successfully DELETED');
+                                    location.reload();
+                                },
+                                error: function () {
+                                    console.log('error')
+                                }
+                            })
+                        }
                     }
+                } else {
+                    alert(`You Don't have a request with id: ${ids}`)
                 }
             },
             error: function (errorThrown) {
@@ -333,7 +345,7 @@ $(window).ready(function () {
 
 
 
-
+//
 $(window).ready(function () {
     $("#change-btn").click(function (e) {
         let ids = $('#cusId').val();
@@ -344,38 +356,40 @@ $(window).ready(function () {
             dataType: "json",
             contentType: "application/json",
             success: function (data) {
-                for (let i = 0; i < data.length; i += 1) {
+                if (myId.includes(ids)) {
+                    for (let i = 0; i < data.length; i += 1) {
 
-                    let newAmount = $('#new-amount').val();
-                    let jsonData = {
-                        "amount": newAmount
+                        let newAmount = $('#new-amount').val();
+                        let jsonData = {
+                            "amount": newAmount
+                        }
+                        if (newAmount == '') {
+                            alert('Incorrect value for new amount')
+                            break;
+                        }
+
+                        if (data[i].id.toString() == ids.toString() && data[i].approval == true) {
+                            alert('Request has already been approved')
+                            i = data.length;
+                            location.reload();
+                        } else if (data[i].id.toString() == ids.toString() && data[i].approval == false) {
+
+                            $.ajax({
+                                type: "PATCH",
+                                url: `http://localhost:3000/borrowers/${ids}`,
+                                dataType: "json",
+                                contentType: 'application/json',
+                                data: JSON.stringify(jsonData),
+                                success: function () {
+                                    alert("Your Request has successfully been modified");
+                                    location.reload();
+                                    // window.location.href = "http://localhost:3000/login.html"
+                                }
+                            })
+                        }
                     }
-                    if (newAmount == '') {
-                        alert('Incorrect value for new amount')
-                        break;
-                    }
-
-                    if (data[i].id.toString() == ids.toString() && data[i].approval == true) {
-                        alert('Request has already been approved')
-                        i = data.length;
-                        location.reload();
-                    } else if (data[i].id.toString() == ids.toString() && data[i].approval == false) {
-
-
-
-                        $.ajax({
-                            type: "PATCH",
-                            url: `http://localhost:3000/borrowers/${ids}`,
-                            dataType: "json",
-                            contentType: 'application/json',
-                            data: JSON.stringify(jsonData),
-                            success: function () {
-                                alert("Your Request has successfully been modified");
-                                location.reload();
-                                // window.location.href = "http://localhost:3000/login.html"
-                            }
-                        })
-                    }
+                } else {
+                    alert(`You Don't have a request with id: ${ids}`)
                 }
             },
             error: function (errorThrown) {
